requirements.txt
tzdata                                 # -> trace
PyYAML
python-dotenv
arrow
openpyxl
openpyxl-stubs
webvtt-py
pymediainfo
mutagen # todo remove (GPLv2+)

torch>=2.4,<2.5

# faster-whisper==1.0.2                  # 2024-05-06 -> faster_whisper_custom
av>=11.0,<13
ctranslate2>=4.0,<5
huggingface_hub>=0.13
tokenizers>=0.13,<1
onnxruntime>=1.14,<2

openai-whisper==20240930                 #
whisper-timestamped==1.15.6              # 2024-04-22: todo remove (GPLv2+)

spylls                                   # 2022-01-23 hunspell compete in Python

spacy>=3.8.2,<3.9
spacy_curated_transformers>=0.2.0,<2     # 2024-09-29 (-> 'de_dep_news_trf') [2.0.0 not compatible to german models]

# stanza
# rapidfuzz
# jaro-winkler
